backend/requirements.txt
flask
flask-cors
scrapy
zenrows
pdfminer.six
python-docx
